+++ packages/next/src/server/dev/hot-reloader-turbopack.ts
@@ -22,6 +22,7 @@ import type {
   WrittenEndpoint,
   TurbopackResult,
   Project,
+  Entrypoints,
 } from '../../build/swc/types'
 import { createDefineEnv } from '../../build/swc'
 import * as Log from '../../build/output/log'
@@ -47,40 +48,31 @@ import {
   AssetMapper,
   type ChangeSubscriptions,
   type ClientState,
-  type EntryIssuesMap,
-  formatIssue,
-  getTurbopackJsConfig,
   handleEntrypoints,
   handlePagesErrorRoute,
   handleRouteType,
   hasEntrypointForKey,
   msToNs,
-  processIssues,
   type ReadyIds,
-  renderStyledStringToErrorAnsi,
   type SendHmr,
   type StartBuilding,
   processTopLevelIssues,
-  type TopLevelIssuesMap,
-  isWellKnownError,
   printNonFatalIssue,
   normalizedPageToTurbopackStructureRoute,
-  isPersistentCachingEnabled,
 } from './turbopack-utils'
 import {
   propagateServerField,
   type ServerFields,
   type SetupOpts,
 } from '../lib/router-utils/setup-dev-bundler'
-import { TurbopackManifestLoader } from './turbopack/manifest-loader'
-import type { Entrypoints } from './turbopack/types'
+import { TurbopackManifestLoader } from '../../shared/lib/turbopack/manifest-loader'
 import { findPagePathData } from './on-demand-entry-handler'
 import type { RouteDefinition } from '../route-definitions/route-definition'
 import {
   type EntryKey,
   getEntryKey,
   splitEntryKey,
-} from './turbopack/entry-key'
+} from '../../shared/lib/turbopack/entry-key'
 import { FAST_REFRESH_RUNTIME_RELOAD } from './messages'
 import { generateEncryptionKeyBase64 } from '../app-render/encryption-utils-server'
 import { isAppPageRouteDefinition } from '../route-definitions/app-page-route-definition'
@@ -92,6 +84,16 @@ import {
   type ModernSourceMapPayload,
 } from '../patch-error-inspect'
 import { getNextErrorFeedbackMiddleware } from '../../client/components/react-dev-overlay/server/get-next-error-feedback-middleware'
+import {
+  formatIssue,
+  getTurbopackJsConfig,
+  isPersistentCachingEnabled,
+  isWellKnownError,
+  processIssues,
+  renderStyledStringToErrorAnsi,
+  type EntryIssuesMap,
+  type TopLevelIssuesMap,
+} from '../../shared/lib/turbopack/utils'
 // import { getSupportedBrowsers } from '../../build/utils'
 
 const wsServer = new ws.Server({ noServer: true })
@@ -973,14 +975,12 @@ export async function createHotReloaderTurbopack(
             let finishBuilding = startBuilding(pathname, requestUrl, false)
             try {
               await handlePagesErrorRoute({
-                dev: true,
                 currentEntryIssues,
                 entrypoints: currentEntrypoints,
                 manifestLoader,
                 devRewrites: opts.fsChecker.rewrites,
                 productionRewrites: undefined,
                 logErrors: true,
-
                 hooks: {
                   subscribeToChanges,
                   handleWrittenEndpoint: (id, result) => {
