+++ packages/next/src/server/dev/turbopack-utils.ts
@@ -1,83 +1,40 @@
-import type { NextConfigComplete } from '../config-shared'
-import loadJsConfig from '../../build/load-jsconfig'
 import type {
   ServerFields,
   SetupOpts,
 } from '../lib/router-utils/setup-dev-bundler'
 import type {
   Issue,
-  StyledString,
   TurbopackResult,
   Endpoint,
-  Entrypoints as RawEntrypoints,
+  RawEntrypoints,
   Update as TurbopackUpdate,
   WrittenEndpoint,
 } from '../../build/swc/types'
-import {
-  decodeMagicIdentifier,
-  MAGIC_IDENTIFIER_REGEX,
-} from '../../shared/lib/magic-identifier'
-import { bold, green, magenta, red } from '../../lib/picocolors'
 import {
   type HMR_ACTION_TYPES,
   HMR_ACTIONS_SENT_TO_BROWSER,
 } from './hot-reloader-types'
 import * as Log from '../../build/output/log'
 import type { PropagateToWorkersField } from '../lib/router-utils/types'
-import type { TurbopackManifestLoader } from './turbopack/manifest-loader'
-import type { AppRoute, Entrypoints, PageRoute } from './turbopack/types'
+import type { TurbopackManifestLoader } from '../../shared/lib/turbopack/manifest-loader'
+import type { AppRoute, Entrypoints, PageRoute } from '../../build/swc/types'
 import {
   type EntryKey,
   getEntryKey,
   splitEntryKey,
-} from './turbopack/entry-key'
+} from '../../shared/lib/turbopack/entry-key'
 import type ws from 'next/dist/compiled/ws'
-import isInternal from '../../shared/lib/is-internal'
 import { isMetadataRoute } from '../../lib/metadata/is-metadata-route'
 import type { CustomRoutes } from '../../lib/load-custom-routes'
-
-export async function getTurbopackJsConfig(
-  dir: string,
-  nextConfig: NextConfigComplete
-) {
-  const { jsConfig } = await loadJsConfig(dir, nextConfig)
-  return jsConfig ?? { compilerOptions: {} }
-}
-
-// An error generated from emitted Turbopack issues. This can include build
-// errors caused by issues with user code.
-export class ModuleBuildError extends Error {
-  name = 'ModuleBuildError'
-}
-
-// An error caused by an internal issue in Turbopack. These should be written
-// to a log file and details should not be shown to the user.
-export class TurbopackInternalError extends Error {
-  name = 'TurbopackInternalError'
-
-  constructor(cause: Error) {
-    super(cause.message)
-    this.stack = cause.stack
-  }
-}
-
-/**
- * Thin stopgap workaround layer to mimic existing wellknown-errors-plugin in webpack's build
- * to emit certain type of errors into cli.
- */
-export function isWellKnownError(issue: Issue): boolean {
-  const { title } = issue
-  const formattedTitle = renderStyledStringToErrorAnsi(title)
-  // TODO: add more well known errors
-  if (
-    formattedTitle.includes('Module not found') ||
-    formattedTitle.includes('Unknown module type')
-  ) {
-    return true
-  }
-
-  return false
-}
+import {
+  formatIssue,
+  getIssueKey,
+  isRelevantWarning,
+  processIssues,
+  renderStyledStringToErrorAnsi,
+  type EntryIssuesMap,
+  type TopLevelIssuesMap,
+} from '../../shared/lib/turbopack/utils'
 
 const onceErrorSet = new Set()
 /**
@@ -116,121 +73,6 @@ export function printNonFatalIssue(issue: Issue) {
   }
 }
 
-function isNodeModulesIssue(issue: Issue): boolean {
-  if (issue.severity === 'warning' && issue.stage === 'config') {
-    // Override for the externalize issue
-    // `Package foo (serverExternalPackages or default list) can't be external`
-    if (
-      renderStyledStringToErrorAnsi(issue.title).includes("can't be external")
-    ) {
-      return false
-    }
-  }
-
-  return (
-    issue.severity === 'warning' &&
-    (issue.filePath.match(/^(?:.*[\\/])?node_modules(?:[\\/].*)?$/) !== null ||
-      // Ignore Next.js itself when running next directly in the monorepo where it is not inside
-      // node_modules anyway.
-      // TODO(mischnic) prevent matches when this is published to npm
-      issue.filePath.startsWith('[project]/packages/next/'))
-  )
-}
-
-export function isRelevantWarning(issue: Issue): boolean {
-  return issue.severity === 'warning' && !isNodeModulesIssue(issue)
-}
-
-export function formatIssue(issue: Issue) {
-  const { filePath, title, description, source } = issue
-  let { documentationLink } = issue
-  let formattedTitle = renderStyledStringToErrorAnsi(title).replace(
-    /\n/g,
-    '\n    '
-  )
-
-  // TODO: Use error codes to identify these
-  // TODO: Generalize adapting Turbopack errors to Next.js errors
-  if (formattedTitle.includes('Module not found')) {
-    // For compatiblity with webpack
-    // TODO: include columns in webpack errors.
-    documentationLink = 'https://nextjs.org/docs/messages/module-not-found'
-  }
-
-  let formattedFilePath = filePath
-    .replace('[project]/', './')
-    .replaceAll('/./', '/')
-    .replace('\\\\?\\', '')
-
-  let message = ''
-
-  if (source && source.range) {
-    const { start } = source.range
-    message = `${formattedFilePath}:${start.line + 1}:${
-      start.column + 1
-    }\n${formattedTitle}`
-  } else if (formattedFilePath) {
-    message = `${formattedFilePath}\n${formattedTitle}`
-  } else {
-    message = formattedTitle
-  }
-  message += '\n'
-
-  if (
-    source?.range &&
-    source.source.content &&
-    // ignore Next.js/React internals, as these can often be huge bundled files.
-    !isInternal(filePath)
-  ) {
-    const { start, end } = source.range
-    const { codeFrameColumns } = require('next/dist/compiled/babel/code-frame')
-
-    message +=
-      codeFrameColumns(
-        source.source.content,
-        {
-          start: {
-            line: start.line + 1,
-            column: start.column + 1,
-          },
-          end: {
-            line: end.line + 1,
-            column: end.column + 1,
-          },
-        },
-        { forceColor: true }
-      ).trim() + '\n\n'
-  }
-
-  if (description) {
-    message += renderStyledStringToErrorAnsi(description) + '\n\n'
-  }
-
-  // TODO: make it possible to enable this for debugging, but not in tests.
-  // if (detail) {
-  //   message += renderStyledStringToErrorAnsi(detail) + '\n\n'
-  // }
-
-  // TODO: Include a trace from the issue.
-
-  if (documentationLink) {
-    message += documentationLink + '\n\n'
-  }
-
-  return message
-}
-
-type IssueKey = `${Issue['severity']}-${Issue['filePath']}-${string}-${string}`
-export type IssuesMap = Map<IssueKey, Issue>
-export type EntryIssuesMap = Map<EntryKey, IssuesMap>
-export type TopLevelIssuesMap = IssuesMap
-
-function getIssueKey(issue: Issue): IssueKey {
-  return `${issue.severity}-${issue.filePath}-${JSON.stringify(
-    issue.title
-  )}-${JSON.stringify(issue.description)}`
-}
-
 export function processTopLevelIssues(
   currentTopLevelIssues: TopLevelIssuesMap,
   result: TurbopackResult
@@ -240,74 +82,6 @@ export function processTopLevelIssues(
   for (const issue of result.issues) {
     const issueKey = getIssueKey(issue)
     currentTopLevelIssues.set(issueKey, issue)
-  }
-}
-
-export function processIssues(
-  currentEntryIssues: EntryIssuesMap,
-  key: EntryKey,
-  result: TurbopackResult,
-  throwIssue: boolean,
-  logErrors: boolean
-) {
-  const newIssues = new Map<IssueKey, Issue>()
-  currentEntryIssues.set(key, newIssues)
-
-  const relevantIssues = new Set()
-
-  for (const issue of result.issues) {
-    if (
-      issue.severity !== 'error' &&
-      issue.severity !== 'fatal' &&
-      issue.severity !== 'warning'
-    )
-      continue
-
-    const issueKey = getIssueKey(issue)
-    newIssues.set(issueKey, issue)
-
-    if (issue.severity !== 'warning') {
-      if (throwIssue) {
-        const formatted = formatIssue(issue)
-        relevantIssues.add(formatted)
-      }
-      // if we throw the issue it will most likely get handed and logged elsewhere
-      else if (logErrors && isWellKnownError(issue)) {
-        const formatted = formatIssue(issue)
-        Log.error(formatted)
-      }
-    }
-  }
-
-  if (relevantIssues.size && throwIssue) {
-    throw new ModuleBuildError([...relevantIssues].join('\n\n'))
-  }
-}
-
-export function renderStyledStringToErrorAnsi(string: StyledString): string {
-  function decodeMagicIdentifiers(str: string): string {
-    return str.replaceAll(MAGIC_IDENTIFIER_REGEX, (ident) => {
-      try {
-        return magenta(`{${decodeMagicIdentifier(ident)}}`)
-      } catch (e) {
-        return magenta(`{${ident} (decoding failed: ${e})}`)
-      }
-    })
-  }
-
-  switch (string.type) {
-    case 'text':
-      return decodeMagicIdentifiers(string.value)
-    case 'strong':
-      return bold(red(decodeMagicIdentifiers(string.value)))
-    case 'code':
-      return green(decodeMagicIdentifiers(string.value))
-    case 'line':
-      return string.value.map(renderStyledStringToErrorAnsi).join('')
-    case 'stack':
-      return string.value.map(renderStyledStringToErrorAnsi).join('\n')
-    default:
-      throw new Error('Unknown StyledString type', string)
   }
 }
 
@@ -789,7 +563,6 @@ export async function handleEntrypoints({
   currentEntryIssues,
   manifestLoader,
   devRewrites,
-  productionRewrites,
   logErrors,
   dev,
 }: {
@@ -803,7 +576,7 @@ export async function handleEntrypoints({
   productionRewrites: CustomRoutes['rewrites'] | undefined
   logErrors: boolean
 
-  dev?: HandleEntrypointsDevOpts
+  dev: HandleEntrypointsDevOpts
 }) {
   currentEntrypoints.global.app = entrypoints.pagesAppEndpoint
   currentEntrypoints.global.document = entrypoints.pagesDocumentEndpoint
@@ -858,12 +631,12 @@ export async function handleEntrypoints({
     // Went from middleware to no middleware
     await dev?.hooks.unsubscribeFromChanges(key)
     currentEntryIssues.delete(key)
-    dev?.hooks.sendHmr('middleware', {
+    dev.hooks.sendHmr('middleware', {
       event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES,
     })
   } else if (!currentEntrypoints.global.middleware && middleware) {
     // Went from no middleware to middleware
-    dev?.hooks.sendHmr('middleware', {
+    dev.hooks.sendHmr('middleware', {
       event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES,
     })
   }
@@ -878,7 +651,7 @@ export async function handleEntrypoints({
       const key = getEntryKey('root', 'server', name)
 
       const writtenEndpoint = await instrumentation[prop].writeToDisk()
-      dev?.hooks.handleWrittenEndpoint(key, writtenEndpoint)
+      dev.hooks.handleWrittenEndpoint(key, writtenEndpoint)
       processIssues(currentEntryIssues, key, writtenEndpoint, false, logErrors)
     }
     await processInstrumentation('instrumentation.nodeJs', 'nodeJs')
@@ -889,25 +662,21 @@ export async function handleEntrypoints({
     )
     await manifestLoader.writeManifests({
       devRewrites,
-      productionRewrites,
+      productionRewrites: undefined,
       entrypoints: currentEntrypoints,
     })
 
-    if (dev) {
-      dev.serverFields.actualInstrumentationHookFile = '/instrumentation'
-      await dev.hooks.propagateServerField(
-        'actualInstrumentationHookFile',
-        dev.serverFields.actualInstrumentationHookFile
-      )
-    }
+    dev.serverFields.actualInstrumentationHookFile = '/instrumentation'
+    await dev.hooks.propagateServerField(
+      'actualInstrumentationHookFile',
+      dev.serverFields.actualInstrumentationHookFile
+    )
   } else {
-    if (dev) {
-      dev.serverFields.actualInstrumentationHookFile = undefined
-      await dev.hooks.propagateServerField(
-        'actualInstrumentationHookFile',
-        dev.serverFields.actualInstrumentationHookFile
-      )
-    }
+    dev.serverFields.actualInstrumentationHookFile = undefined
+    await dev.hooks.propagateServerField(
+      'actualInstrumentationHookFile',
+      dev.serverFields.actualInstrumentationHookFile
+    )
   }
 
   if (middleware) {
@@ -917,7 +686,7 @@ export async function handleEntrypoints({
 
     async function processMiddleware() {
       const writtenEndpoint = await endpoint.writeToDisk()
-      dev?.hooks.handleWrittenEndpoint(key, writtenEndpoint)
+      dev.hooks.handleWrittenEndpoint(key, writtenEndpoint)
       processIssues(currentEntryIssues, key, writtenEndpoint, false, logErrors)
       await manifestLoader.loadMiddlewareManifest('middleware', 'middleware')
       if (dev) {
@@ -953,7 +722,7 @@ export async function handleEntrypoints({
           )
           await manifestLoader.writeManifests({
             devRewrites,
-            productionRewrites,
+            productionRewrites: undefined,
             entrypoints: currentEntrypoints,
           })
 
@@ -971,22 +740,18 @@ export async function handleEntrypoints({
     manifestLoader.deleteMiddlewareManifest(
       getEntryKey('root', 'server', 'middleware')
     )
-    if (dev) {
-      dev.serverFields.actualMiddlewareFile = undefined
-      dev.serverFields.middleware = undefined
-    }
+    dev.serverFields.actualMiddlewareFile = undefined
+    dev.serverFields.middleware = undefined
   }
 
-  if (dev) {
-    await dev.hooks.propagateServerField(
-      'actualMiddlewareFile',
-      dev.serverFields.actualMiddlewareFile
-    )
-    await dev.hooks.propagateServerField(
-      'middleware',
-      dev.serverFields.middleware
-    )
-  }
+  await dev.hooks.propagateServerField(
+    'actualMiddlewareFile',
+    dev.serverFields.actualMiddlewareFile
+  )
+  await dev.hooks.propagateServerField(
+    'middleware',
+    dev.serverFields.middleware
+  )
 }
 
 async function handleEntrypointsDevCleanup({
@@ -1050,49 +815,43 @@ async function handleEntrypointsDevCleanup({
 }
 
 export async function handlePagesErrorRoute({
-  dev,
   currentEntryIssues,
   entrypoints,
   manifestLoader,
   devRewrites,
   productionRewrites,
   logErrors,
-
   hooks,
 }: {
-  dev: boolean
   currentEntryIssues: EntryIssuesMap
   entrypoints: Entrypoints
   manifestLoader: TurbopackManifestLoader
   devRewrites: SetupOpts['fsChecker']['rewrites'] | undefined
   productionRewrites: CustomRoutes['rewrites'] | undefined
   logErrors: boolean
-
-  hooks?: HandleRouteTypeHooks // dev
+  hooks: HandleRouteTypeHooks
 }) {
   if (entrypoints.global.app) {
     const key = getEntryKey('pages', 'server', '_app')
 
     const writtenEndpoint = await entrypoints.global.app.writeToDisk()
-    hooks?.handleWrittenEndpoint(key, writtenEndpoint)
-    if (dev) {
-      hooks?.subscribeToChanges(
-        key,
-        false,
-        entrypoints.global.app,
-        () => {
-          // There's a special case for this in `../client/page-bootstrap.ts`.
-          // https://github.com/vercel/next.js/blob/08d7a7e5189a835f5dcb82af026174e587575c0e/packages/next/src/client/page-bootstrap.ts#L69-L71
-          return { event: HMR_ACTIONS_SENT_TO_BROWSER.CLIENT_CHANGES }
-        },
-        () => {
-          return {
-            action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
-            data: '_app has changed (error route)',
-          }
+    hooks.handleWrittenEndpoint(key, writtenEndpoint)
+    hooks.subscribeToChanges(
+      key,
+      false,
+      entrypoints.global.app,
+      () => {
+        // There's a special case for this in `../client/page-bootstrap.ts`.
+        // https://github.com/vercel/next.js/blob/08d7a7e5189a835f5dcb82af026174e587575c0e/packages/next/src/client/page-bootstrap.ts#L69-L71
+        return { event: HMR_ACTIONS_SENT_TO_BROWSER.CLIENT_CHANGES }
+      },
+      () => {
+        return {
+          action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
+          data: '_app has changed (error route)',
         }
-      )
-    }
+      }
+    )
     processIssues(currentEntryIssues, key, writtenEndpoint, false, logErrors)
   }
   await manifestLoader.loadBuildManifest('_app')
@@ -1103,26 +862,24 @@ export async function handlePagesErrorRoute({
     const key = getEntryKey('pages', 'server', '_document')
 
     const writtenEndpoint = await entrypoints.global.document.writeToDisk()
-    hooks?.handleWrittenEndpoint(key, writtenEndpoint)
-    if (dev) {
-      hooks?.subscribeToChanges(
-        key,
-        false,
-        entrypoints.global.document,
-        () => {
-          return {
-            action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
-            data: '_document has changed (error route)',
-          }
-        },
-        (e) => {
-          return {
-            action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
-            data: `error in _document subscription (error route): ${e}`,
-          }
+    hooks.handleWrittenEndpoint(key, writtenEndpoint)
+    hooks.subscribeToChanges(
+      key,
+      false,
+      entrypoints.global.document,
+      () => {
+        return {
+          action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
+          data: '_document has changed (error route)',
         }
-      )
-    }
+      },
+      (e) => {
+        return {
+          action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
+          data: `error in _document subscription (error route): ${e}`,
+        }
+      }
+    )
     processIssues(currentEntryIssues, key, writtenEndpoint, false, logErrors)
   }
   await manifestLoader.loadPagesManifest('_document')
@@ -1131,25 +888,23 @@ export async function handlePagesErrorRoute({
     const key = getEntryKey('pages', 'server', '_error')
 
     const writtenEndpoint = await entrypoints.global.error.writeToDisk()
-    hooks?.handleWrittenEndpoint(key, writtenEndpoint)
-    if (dev) {
-      hooks?.subscribeToChanges(
-        key,
-        false,
-        entrypoints.global.error,
-        () => {
-          // There's a special case for this in `../client/page-bootstrap.ts`.
-          // https://github.com/vercel/next.js/blob/08d7a7e5189a835f5dcb82af026174e587575c0e/packages/next/src/client/page-bootstrap.ts#L69-L71
-          return { event: HMR_ACTIONS_SENT_TO_BROWSER.CLIENT_CHANGES }
-        },
-        (e) => {
-          return {
-            action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
-            data: `error in _error subscription: ${e}`,
-          }
+    hooks.handleWrittenEndpoint(key, writtenEndpoint)
+    hooks.subscribeToChanges(
+      key,
+      false,
+      entrypoints.global.error,
+      () => {
+        // There's a special case for this in `../client/page-bootstrap.ts`.
+        // https://github.com/vercel/next.js/blob/08d7a7e5189a835f5dcb82af026174e587575c0e/packages/next/src/client/page-bootstrap.ts#L69-L71
+        return { event: HMR_ACTIONS_SENT_TO_BROWSER.CLIENT_CHANGES }
+      },
+      (e) => {
+        return {
+          action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
+          data: `error in _error subscription: ${e}`,
         }
-      )
-    }
+      }
+    )
     processIssues(currentEntryIssues, key, writtenEndpoint, false, logErrors)
   }
   await manifestLoader.loadBuildManifest('_error')
@@ -1201,10 +956,4 @@ export function normalizedPageToTurbopackStructureRoute(
     entrypointKey = entrypointKey + '/route'
   }
   return entrypointKey
-}
-
-export function isPersistentCachingEnabled(
-  config: NextConfigComplete
-): boolean {
-  return config.experimental.turbo?.unstablePersistentCaching || false
 }
